Separates GroupCollectionDiff tests from CountScenario tests
