+++ tools/tck-inspection/src/test/scala/org/opencypher/tools/tck/inspection/CountScenariosTest.scala
@@ -44,15 +44,15 @@ import org.scalatest.FunSuite
 import org.scalatest.Matchers
 
 class CountScenariosTest extends FunSuite with Matchers {
-  val dummyPickle = new Pickle("", "", new util.ArrayList[PickleStep](), new util.ArrayList[PickleTag](), new util.ArrayList[PickleLocation]())
-  val dummyPickleStep = new PickleStep("", new util.ArrayList[Argument](), new util.ArrayList[PickleLocation]())
-  def namedDummyPickleStep(name: String) = new PickleStep(name, new util.ArrayList[Argument](), new util.ArrayList[PickleLocation]())
-  def dummyStep(name: String): Step = Dummy(namedDummyPickleStep(name))
-  def dummyPath(path: String): java.nio.file.Path = new java.io.File("ftr1.feature").toPath
+  private val dummyPickle = new Pickle("", "", new util.ArrayList[PickleStep](), new util.ArrayList[PickleTag](), new util.ArrayList[PickleLocation]())
+  private val dummyPickleStep = new PickleStep("", new util.ArrayList[Argument](), new util.ArrayList[PickleLocation]())
+  private def namedDummyPickleStep(name: String) = new PickleStep(name, new util.ArrayList[Argument](), new util.ArrayList[PickleLocation]())
+  private def dummyStep(name: String): Step = Dummy(namedDummyPickleStep(name))
+  private def dummyPath(path: String): java.nio.file.Path = new java.io.File("ftr1.feature").toPath
 
   test("Count single top-level scenario without tags") {
     val scenarios: Seq[Scenario] = Seq(Scenario(List[String](), "ftr1", "scr1", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature")))
-    val expectedCountOutput =
+    val expectedCountOutput: String =
       """Total                 1
         || Feature: ftr1       1""".stripMargin
     CountScenarios.reportCountsInPrettyPrint(GroupCollection(scenarios)) should equal(expectedCountOutput)
@@ -60,7 +60,7 @@ class CountScenariosTest extends FunSuite with Matchers {
 
   test("Count single top-level scenario with tag") {
     val scenarios: Seq[Scenario] = Seq(Scenario(List[String](), "ftr1", "scr1", None, Set[String]("A", "@B", "C"), List[Step](), dummyPickle, dummyPath("ftr1.feature")))
-    val expectedCountOutput =
+    val expectedCountOutput: String =
       """Total                 1
         || Feature: ftr1       1
         || @B                  1
@@ -71,7 +71,7 @@ class CountScenariosTest extends FunSuite with Matchers {
 
   test("Count single sub-level scenario without tags") {
     val scenarios: Seq[Scenario] = Seq(Scenario(List[String]("A", "B", "C"), "ftr1", "scr1", None, Set[String](), List[Step](), dummyPickle, dummyPath("A/B/C/ftr1.feature")))
-    val expectedCountOutput =
+    val expectedCountOutput: String =
       """Total                       1
         || A                         1
         || | B                       1
@@ -84,7 +84,7 @@ class CountScenariosTest extends FunSuite with Matchers {
     val scenarios: Seq[Scenario] = Seq(
       Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature")),
       Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature")))
-    val expectedCountOutput =
+    val expectedCountOutput: String =
       """Total                 2
         || Feature: ftr1       2""".stripMargin
     CountScenarios.reportCountsInPrettyPrint(GroupCollection(scenarios)) should equal(expectedCountOutput)
@@ -95,7 +95,7 @@ class CountScenariosTest extends FunSuite with Matchers {
       Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature")),
       Scenario(List[String](), "ftr2", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr2.feature")),
       Scenario(List[String](), "ftr1", "scrC", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature")))
-    val expectedCountOutput =
+    val expectedCountOutput: String =
       """Total                 3
         || Feature: ftr1       2
         || Feature: ftr2       1""".stripMargin
@@ -107,7 +107,7 @@ class CountScenariosTest extends FunSuite with Matchers {
       Scenario(List[String](), "ftr1", "scrA", None, Set[String]("C", "B"), List[Step](), dummyPickle, dummyPath("ftr1.feature")),
       Scenario(List[String](), "ftr2", "scrB", None, Set[String]("C", "D", "A"), List[Step](), dummyPickle, dummyPath("ftr2.feature")),
       Scenario(List[String](), "ftr1", "scrC", None, Set[String]("A", "C"), List[Step](), dummyPickle, dummyPath("ftr1.feature")))
-    val expectedCountOutput =
+    val expectedCountOutput: String =
       """Total                 3
         || Feature: ftr1       2
         || Feature: ftr2       1
@@ -123,7 +123,7 @@ class CountScenariosTest extends FunSuite with Matchers {
       Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature")),
       Scenario(List[String]("T", "C", "K"), "ftr2", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("T/C/K/ftr2.feature")),
       Scenario(List[String]("T", "A", "K"), "ftr1", "scrC", None, Set[String](), List[Step](), dummyPickle, dummyPath("T/C/A/ftr1.feature")))
-    val expectedCountOutput =
+    val expectedCountOutput: String =
       """Total                       3
         || T                         2
         || | A                       1
@@ -143,7 +143,7 @@ class CountScenariosTest extends FunSuite with Matchers {
       Scenario(List[String]("T", "A", "K"), "ftr1", "scrC", None, Set[String]("A", "C"), List[Step](), dummyPickle, dummyPath("T/A/K/ftr1.feature")),
       Scenario(List[String]("T", "C", "K"), "ftr2", "scrD", None, Set[String](), List[Step](), dummyPickle, dummyPath("T/C/K/ftr2.feature")),
       Scenario(List[String]("T"), "ftr3", "scrE", None, Set[String]("B", "A", "C"), List[Step](), dummyPickle, dummyPath("T/ftr3.feature")))
-    val expectedCountOutput =
+    val expectedCountOutput: String =
       """Total                       5
         || T                         4
         || | A                       1
@@ -164,7 +164,7 @@ class CountScenariosTest extends FunSuite with Matchers {
   test("Count scenarios through TCK API") {
     val fooUri: URI = getClass.getResource("cucumber").toURI
     val scenarios: Seq[Scenario] = CypherTCK.parseFeatures(fooUri).flatMap(_.scenarios)
-    val expectedCountOutput =
+    val expectedCountOutput: String =
       """Total                        13
         || foo                         8
         || | bar                       6
@@ -182,145 +182,69 @@ class CountScenariosTest extends FunSuite with Matchers {
     CountScenarios.reportCountsInPrettyPrint(GroupCollection(scenarios)) should equal(expectedCountOutput)
   }
 
-  test("Diff with one scenario added from a top-level same feature without tags") {
-    val scrA = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA, scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
-
-    val expectedResult = Map[Group, GroupDiff](
-      Total -> GroupDiff(Set(scrB), Set(), Set(), Set(scrA), Set()),
-      Feature("ftr1", 1, Some(Total)) -> GroupDiff(Set(scrB), Set(), Set(), Set(scrA), Set())
-    )
-
-    GroupCollectionDiff(collectBefore, collectAfter) should equal(expectedResult)
-  }
-
   test("Report pretty diff counts with one scenario added from a top-level same feature without tags") {
     val scrA = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
     val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA, scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
+    val groupCollectionDiff: Map[Group, GroupDiff] = GroupCollectionDiff(
+      GroupCollection(Seq(scrB)),
+      GroupCollection(Seq(scrA, scrB))
+    )
 
-    val expectedResult =
+    val expectedResult: String =
       """Group           unchanged  moved only  changed more  added  removed
         |–––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
         |Total                   1           0             0      1        0
         |- Feature: ftr1         1           0             0      1        0""".stripMargin
 
-    CountScenarios.reportDiffCountsInPrettyPrint(GroupCollectionDiff(collectBefore, collectAfter)) should equal(expectedResult)
-  }
-
-  test("Diff with one scenario removed from a top-level same feature without tags") {
-    val scrA = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA, scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
-
-    val expectedResult = Map[Group, GroupDiff](
-      Total -> GroupDiff(Set(scrB), Set(), Set(), Set(), Set(scrA)),
-      Feature("ftr1", 1, Some(Total)) -> GroupDiff(Set(scrB), Set(), Set(), Set(), Set(scrA))
-    )
-
-    GroupCollectionDiff(collectBefore, collectAfter) should equal(expectedResult)
+    CountScenarios.reportDiffCountsInPrettyPrint(groupCollectionDiff) should equal(expectedResult)
   }
 
   test("Report pretty diff counts with one scenario removed from a top-level same feature without tags") {
     val scrA = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
     val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA, scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
+    val groupCollectionDiff: Map[Group, GroupDiff] = GroupCollectionDiff(
+      GroupCollection(Seq(scrA, scrB)),
+      GroupCollection(Seq(scrB))
+    )
 
-    val expectedResult =
+    val expectedResult: String =
       """Group           unchanged  moved only  changed more  added  removed
         |–––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
         |Total                   1           0             0      0        1
         |- Feature: ftr1         1           0             0      0        1""".stripMargin
 
-    CountScenarios.reportDiffCountsInPrettyPrint(GroupCollectionDiff(collectBefore, collectAfter)) should equal(expectedResult)
-  }
-
-  test("Diff with one scenario moved to another top-level feature without tags") {
-    val scrA1 = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scrA2 = Scenario(List[String](), "ftr2", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr2.feature"))
-    val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA1, scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA2, scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
-
-    val scrA1scrA2Diff = ScenarioDiff(scrA1, scrA2)
-    scrA1scrA2Diff.diffTags should equal(Set(Moved))
-
-    val expectedResult = Map[Group, GroupDiff](
-      Total -> GroupDiff(Set(scrB), Set((scrA1, scrA2, scrA1scrA2Diff)), Set(), Set(), Set()),
-      Feature("ftr1", 1, Some(Total)) -> GroupDiff(Set(scrB), Set(), Set(), Set(), Set(scrA1)),
-      Feature("ftr2", 1, Some(Total)) -> GroupDiff(Set(), Set(), Set(), Set(scrA2), Set())
-    )
-
-    GroupCollectionDiff(collectBefore, collectAfter) should equal(expectedResult)
+    CountScenarios.reportDiffCountsInPrettyPrint(groupCollectionDiff) should equal(expectedResult)
   }
 
   test("Report pretty diff counts with one scenario moved to another top-level feature without tags") {
     val scrA1 = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
     val scrA2 = Scenario(List[String](), "ftr2", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr2.feature"))
     val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA1, scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA2, scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
+    val groupCollectionDiff: Map[Group, GroupDiff] = GroupCollectionDiff(
+      GroupCollection(Seq(scrA1, scrB)),
+      GroupCollection(Seq(scrA2, scrB))
+    )
 
-    val expectedResult =
+    val expectedResult: String =
       """Group           unchanged  moved only  changed more  added  removed
         |–––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
         |Total                   1           1             0      0        0
         |- Feature: ftr1         1           0             0      0        1
         |- Feature: ftr2         0           0             0      1        0""".stripMargin
 
-    CountScenarios.reportDiffCountsInPrettyPrint(GroupCollectionDiff(collectBefore, collectAfter)) should equal(expectedResult)
-  }
-
-  test("Diff with one scenario moved to another sub-level feature without tags") {
-    val scrA1 = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scrA2 = Scenario(List[String]("X"), "ftr2", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("X/ftr2.feature"))
-    val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA1, scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA2, scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
-
-    val scrA1scrA2Diff = ScenarioDiff(scrA1, scrA2)
-    scrA1scrA2Diff.diffTags should equal(Set(Moved))
-
-    val catX = ScenarioCategory("X", 1, Some(Total))
-    val expectedResult = Map[Group, GroupDiff](
-      Total -> GroupDiff(Set(scrB), Set((scrA1, scrA2, scrA1scrA2Diff)), Set(), Set(), Set()),
-      Feature("ftr1", 1, Some(Total)) -> GroupDiff(Set(scrB), Set(), Set(), Set(), Set(scrA1)),
-      catX -> GroupDiff(Set(), Set(), Set(), Set(scrA2), Set()),
-      Feature("ftr2", 2, Some(catX)) -> GroupDiff(Set(), Set(), Set(), Set(scrA2), Set())
-    )
-
-    GroupCollectionDiff(collectBefore, collectAfter) should equal(expectedResult)
+    CountScenarios.reportDiffCountsInPrettyPrint(groupCollectionDiff) should equal(expectedResult)
   }
 
   test("Report pretty diff counts with one scenario moved to another sub-level feature without tags") {
     val scrA1 = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
     val scrA2 = Scenario(List[String]("X"), "ftr2", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("X/ftr2.feature"))
     val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA1, scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA2, scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
+    val groupCollectionDiff: Map[Group, GroupDiff] = GroupCollectionDiff(
+      GroupCollection(Seq(scrA1, scrB)),
+      GroupCollection(Seq(scrA2, scrB))
+    )
 
-    val expectedResult =
+    val expectedResult: String =
       """Group             unchanged  moved only  changed more  added  removed
         |–––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
         |Total                     1           1             0      0        0
@@ -328,33 +252,7 @@ class CountScenariosTest extends FunSuite with Matchers {
         |  - Feature: ftr2         0           0             0      1        0
         |- Feature: ftr1           1           0             0      0        1""".stripMargin
 
-    CountScenarios.reportDiffCountsInPrettyPrint(GroupCollectionDiff(collectBefore, collectAfter)) should equal(expectedResult)
-  }
-
-  test("Diff with one scenario moved to another top-level feature and a changed tag") {
-    val scrA1 = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](dummyStep("A")), dummyPickle, dummyPath("ftr1.feature"))
-    val scrA2 = Scenario(List[String](), "ftr2", "scrA", None, Set[String](), List[Step](dummyStep("A")), dummyPickle, dummyPath("ftr2.feature"))
-    val scrB1 = Scenario(List[String](), "ftr1", "scrB", None, Set[String]("A"), List[Step](dummyStep("B")), dummyPickle, dummyPath("ftr1.feature"))
-    val scrB2 = Scenario(List[String](), "ftr1", "scrB", None, Set[String]("B"), List[Step](dummyStep("B")), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA1, scrB1)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA2, scrB2)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
-
-    val scrA1scrA2Diff = ScenarioDiff(scrA1, scrA2)
-    scrA1scrA2Diff.diffTags should equal(Set(Moved))
-    val scrB1scrB2Diff = ScenarioDiff(scrB1, scrB2)
-    scrB1scrB2Diff.diffTags should equal(Set(Retagged))
-
-    val expectedResult = Map[Group, GroupDiff](
-      Total -> GroupDiff(Set(), Set((scrA1, scrA2, scrA1scrA2Diff)), Set((scrB1, scrB2, scrB1scrB2Diff)), Set(), Set()),
-      Feature("ftr1", 1, Some(Total)) -> GroupDiff(Set(), Set(), Set((scrB1, scrB2, scrB1scrB2Diff)), Set(), Set(scrA1)),
-      Feature("ftr2", 1, Some(Total)) -> GroupDiff(Set(), Set(), Set(), Set(scrA2), Set()),
-      Tag("A") -> GroupDiff(Set(), Set(), Set(), Set(), Set(scrB1)),
-      Tag("B") -> GroupDiff(Set(), Set(), Set(), Set(scrB2), Set())
-    )
-
-    GroupCollectionDiff(collectBefore, collectAfter) should equal(expectedResult)
+    CountScenarios.reportDiffCountsInPrettyPrint(groupCollectionDiff) should equal(expectedResult)
   }
 
   test("Report pretty diff counts with one scenario moved to another top-level feature and a changed tag") {
@@ -362,12 +260,12 @@ class CountScenariosTest extends FunSuite with Matchers {
     val scrA2 = Scenario(List[String](), "ftr2", "scrA", None, Set[String](), List[Step](dummyStep("A")), dummyPickle, dummyPath("ftr2.feature"))
     val scrB1 = Scenario(List[String](), "ftr1", "scrB", None, Set[String]("A"), List[Step](dummyStep("B")), dummyPickle, dummyPath("ftr1.feature"))
     val scrB2 = Scenario(List[String](), "ftr1", "scrB", None, Set[String]("B"), List[Step](dummyStep("B")), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA1, scrB1)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA2, scrB2)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
+    val groupCollectionDiff: Map[Group, GroupDiff] = GroupCollectionDiff(
+      GroupCollection(Seq(scrA1, scrB1)),
+      GroupCollection(Seq(scrA2, scrB2))
+    )
 
-    val expectedResult =
+    val expectedResult: String =
       """Group           unchanged  moved only  changed more  added  removed
         |–––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
         |Total                   0           1             1      0        0
@@ -376,31 +274,7 @@ class CountScenariosTest extends FunSuite with Matchers {
         |- A                     0           0             0      0        1
         |- B                     0           0             0      1        0""".stripMargin
 
-    CountScenarios.reportDiffCountsInPrettyPrint(GroupCollectionDiff(collectBefore, collectAfter)) should equal(expectedResult)
-  }
-
-  test("Diff with one scenario from a outline in a top-level feature has a changed tags") {
-    val scrA = Scenario(List[String](), "ftr1", "scrA", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scrB0 = Scenario(List[String](), "ftr1", "scrB", Some(0), Set[String]("A"), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scrB1 = Scenario(List[String](), "ftr1", "scrB", Some(1), Set[String]("A"), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scrB2 = Scenario(List[String](), "ftr1", "scrB", Some(2), Set[String]("A"), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scrB1x = Scenario(List[String](), "ftr1", "scrB", Some(1), Set[String]("B"), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA, scrB0, scrB1, scrB2)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA, scrB0, scrB1x, scrB2)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
-
-    val scrB1scrB1xDiff = ScenarioDiff(scrB1, scrB1x)
-    scrB1scrB1xDiff.diffTags should equal(Set(Retagged))
-
-    val expectedResult = Map[Group, GroupDiff](
-      Total -> GroupDiff(Set(scrA, scrB0, scrB2), Set(), Set((scrB1, scrB1x, scrB1scrB1xDiff)), Set(), Set()),
-      Feature("ftr1", 1, Some(Total)) -> GroupDiff(Set(scrA, scrB0, scrB2), Set(), Set((scrB1, scrB1x, scrB1scrB1xDiff)), Set(), Set()),
-      Tag("A") -> GroupDiff(Set(scrB0, scrB2), Set(), Set(), Set(), Set(scrB1)),
-      Tag("B") -> GroupDiff(Set(), Set(), Set(), Set(scrB1x), Set())
-    )
-
-    GroupCollectionDiff(collectBefore, collectAfter) should equal(expectedResult)
+    CountScenarios.reportDiffCountsInPrettyPrint(groupCollectionDiff) should equal(expectedResult)
   }
 
   test("Report pretty diff counts with one scenario from a outline in a top-level feature has a changed tags") {
@@ -409,12 +283,12 @@ class CountScenariosTest extends FunSuite with Matchers {
     val scrB1 = Scenario(List[String](), "ftr1", "scrB", Some(1), Set[String]("A"), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
     val scrB2 = Scenario(List[String](), "ftr1", "scrB", Some(2), Set[String]("A"), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
     val scrB1x = Scenario(List[String](), "ftr1", "scrB", Some(1), Set[String]("B"), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA, scrB0, scrB1, scrB2)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA, scrB0, scrB1x, scrB2)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
+    val groupCollectionDiff: Map[Group, GroupDiff] = GroupCollectionDiff(
+      GroupCollection(Seq(scrA, scrB0, scrB1, scrB2)),
+      GroupCollection(Seq(scrA, scrB0, scrB1x, scrB2))
+    )
 
-    val expectedResult =
+    val expectedResult: String =
       """Group           unchanged  moved only  changed more  added  removed
         |–––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
         |Total                   3           0             1      0        0
@@ -422,46 +296,21 @@ class CountScenariosTest extends FunSuite with Matchers {
         |- A                     2           0             0      0        1
         |- B                     0           0             0      1        0""".stripMargin
 
-    CountScenarios.reportDiffCountsInPrettyPrint(GroupCollectionDiff(collectBefore, collectAfter)) should equal(expectedResult)
-  }
-
-  test("Diff with one scenario changed in categories, tags, and content of steps") {
-    val stepsA1 = List[Step](Dummy(dummyPickleStep), Measure(dummyPickleStep))
-    val stepsA2 = List[Step](Measure(dummyPickleStep), Dummy(dummyPickleStep))
-    val scrA1 = Scenario(List[String](), "ftr1", "scrA", None, Set[String]("T"), stepsA1, dummyPickle, dummyPath("ftr1.feature"))
-    val scrA2 = Scenario(List[String](), "ftr2", "scrA", None, Set[String]("X"), stepsA2, dummyPickle, dummyPath("ftr2.feature"))
-    val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA1, scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA2, scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
-
-    val scrA1scrA2Diff = ScenarioDiff(scrA1, scrA2)
-    scrA1scrA2Diff.diffTags should equal(Set(Moved, Retagged, StepsChanged))
-
-    val expectedResult = Map[Group, GroupDiff](
-      Total -> GroupDiff(Set(scrB), Set(), Set((scrA1, scrA2, scrA1scrA2Diff)), Set(), Set()),
-      Feature("ftr1", 1, Some(Total)) -> GroupDiff(Set(scrB), Set(), Set(), Set(), Set(scrA1)),
-      Feature("ftr2", 1, Some(Total)) -> GroupDiff(Set(), Set(), Set(), Set(scrA2), Set()),
-      Tag("T") -> GroupDiff(Set(), Set(), Set(), Set(), Set(scrA1)),
-      Tag("X") -> GroupDiff(Set(), Set(), Set(), Set(scrA2), Set())
-    )
-
-    GroupCollectionDiff(collectBefore, collectAfter) should equal(expectedResult)
+    CountScenarios.reportDiffCountsInPrettyPrint(groupCollectionDiff) should equal(expectedResult)
   }
 
   test("Report pretty diff counts with one scenario changed in categories, tags, and content of steps") {
-    val stepsA1 = List[Step](Dummy(dummyPickleStep), Measure(dummyPickleStep))
-    val stepsA2 = List[Step](Measure(dummyPickleStep), Dummy(dummyPickleStep))
+    val stepsA1: List[Step] = List[Step](Dummy(dummyPickleStep), Measure(dummyPickleStep))
+    val stepsA2: List[Step] = List[Step](Measure(dummyPickleStep), Dummy(dummyPickleStep))
     val scrA1 = Scenario(List[String](), "ftr1", "scrA", None, Set[String]("T"), stepsA1, dummyPickle, dummyPath("ftr1.feature"))
     val scrA2 = Scenario(List[String](), "ftr2", "scrA", None, Set[String]("X"), stepsA2, dummyPickle, dummyPath("ftr2.feature"))
     val scrB = Scenario(List[String](), "ftr1", "scrB", None, Set[String](), List[Step](), dummyPickle, dummyPath("ftr1.feature"))
-    val scenariosBefore: Seq[Scenario] = Seq(scrA1, scrB)
-    val scenariosAfter: Seq[Scenario] = Seq(scrA2, scrB)
-    val collectBefore = GroupCollection(scenariosBefore)
-    val collectAfter = GroupCollection(scenariosAfter)
+    val groupCollectionDiff: Map[Group, GroupDiff] = GroupCollectionDiff(
+      GroupCollection(Seq(scrA1, scrB)),
+      GroupCollection(Seq(scrA2, scrB))
+    )
 
-    val expectedResult =
+    val expectedResult: String =
       """Group           unchanged  moved only  changed more  added  removed
         |–––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
         |Total                   1           0             1      0        0
@@ -470,7 +319,7 @@ class CountScenariosTest extends FunSuite with Matchers {
         |- T                     0           0             0      0        1
         |- X                     0           0             0      1        0""".stripMargin
 
-    CountScenarios.reportDiffCountsInPrettyPrint(GroupCollectionDiff(collectBefore, collectAfter)) should equal(expectedResult)
+    CountScenarios.reportDiffCountsInPrettyPrint(groupCollectionDiff) should equal(expectedResult)
   }
 
   test("Diff scenarios through TCK API") {
@@ -478,7 +327,7 @@ class CountScenariosTest extends FunSuite with Matchers {
     val fooUriAfter: URI = getClass.getResource("cucumberDiff").toURI
     val scenariosBefore: Seq[Scenario] = CypherTCK.parseFeatures(fooUriBefore).flatMap(_.scenarios)
     val scenariosAfter: Seq[Scenario] = CypherTCK.parseFeatures(fooUriAfter).flatMap(_.scenarios)
-    val expectedResult =
+    val expectedResult: String =
       """Group                   unchanged  moved only  changed more  added  removed
         |–––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
         |Total                          12           1             0      0        0
